week_10/day_3/day_3/abstract_classes/code/Airline/src/main/java/Pilot.java
public class Pilot extends AbstractPerson {

    private String licenseNumber;
    private String rank;
    public Pilot(String _name, String _rank, String _licenseNumber){
        super(_name);
        this.rank = _rank;
        this.licenseNumber = _licenseNumber;
    }



    public String getLicenseNumber() {
        return licenseNumber;
    }
    public void setLicenseNumber(String licenseNumber) {
        this.licenseNumber = licenseNumber;
    }
    public String getRank() {
        return rank;
    }
    public void setRank(String rank) {
        this.rank = rank;
    }

    public String flyPlane() {
        return "The Pilot is now flying this plane";
    }
}
